requirements.txt
pydantic>=2.7
jsonschema>=4.22
tqdm>=4.66
python-dotenv>=1.0
typer>=0.12
rich>=13.7
orjson>=3.10
regex>=2024.4.28
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
httpx>=0.27
tenacity>=8.3
